photo modals ultra refactor
