+++ frontend/src/components/Grievances/GrievanceFlexFieldPhotoModal.tsx
@@ -1,18 +1,11 @@
-import { Box, Button, DialogContent } from '@material-ui/core';
-import React, { useState } from 'react';
+import { Box } from '@material-ui/core';
+import React from 'react';
 import { useParams } from 'react-router-dom';
-import { Dialog } from '../../containers/dialogs/Dialog';
-import { DialogActions } from '../../containers/dialogs/DialogActions';
 import {
   AllAddIndividualFieldsQuery,
   useGrievanceTicketFlexFieldsQuery,
 } from '../../__generated__/graphql';
-import { PhotoModalHeader } from '../PhotoModal/PhotoModalHeader';
-import {
-  DialogFooter,
-  MiniImage,
-  StyledImage,
-} from '../population/IndividualFlexFieldPhotoModal';
+import { PhotoModal } from '../PhotoModal/PhotoModal';
 
 export interface GrievanceFlexFieldPhotoModalProps {
   field: AllAddIndividualFieldsQuery['allAddIndividualsFieldsAttributes'][number];
@@ -26,12 +19,10 @@ export const GrievanceFlexFieldPhotoModal = ({
   isIndividual,
 }: GrievanceFlexFieldPhotoModalProps): React.ReactElement => {
   const { id } = useParams();
-  const [turnAngle, setTurnAngle] = useState(90);
   const { data } = useGrievanceTicketFlexFieldsQuery({
     variables: { id },
     fetchPolicy: 'network-only',
   });
-  const [dialogOpen, setDialogOpen] = useState(false);
   if (!data) {
     return null;
   }
@@ -46,26 +37,7 @@ export const GrievanceFlexFieldPhotoModal = ({
     ? flexFields[field.name]?.previous_value
     : flexFields[field.name]?.value;
   return picUrl ? (
-    <>
-      <MiniImage alt='photo' src={picUrl} onClick={() => setDialogOpen(true)} />
-      <Dialog
-        open={dialogOpen}
-        onClose={() => setDialogOpen(false)}
-        aria-labelledby='form-dialog-title'
-      >
-        <PhotoModalHeader turnAngle={turnAngle} setTurnAngle={setTurnAngle} />
-        <DialogContent>
-          <Box p={3}>
-            <StyledImage id='modalImg' alt='photo' src={picUrl} />
-          </Box>
-        </DialogContent>
-        <DialogFooter>
-          <DialogActions>
-            <Button onClick={() => setDialogOpen(false)}>CANCEL</Button>
-          </DialogActions>
-        </DialogFooter>
-      </Dialog>
-    </>
+    <PhotoModal src={picUrl} />
   ) : (
     <Box style={{ height: '100%' }} display='flex' alignItems='center'>
       -
