+++ frontend/src/components/Grievances/GrievanceFlexFieldPhotoModalEditable.tsx
@@ -1,20 +1,12 @@
-import { Box, Button, DialogContent, IconButton } from '@material-ui/core';
-import CloseIcon from '@material-ui/icons/Close';
+import { Box } from '@material-ui/core';
 import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { Dialog } from '../../containers/dialogs/Dialog';
-import { DialogActions } from '../../containers/dialogs/DialogActions';
 import { FormikFileField } from '../../shared/Formik/FormikFileField';
 import {
   AllAddIndividualFieldsQuery,
   useGrievanceTicketFlexFieldsQuery,
 } from '../../__generated__/graphql';
-import { PhotoModalHeader } from '../PhotoModal/PhotoModalHeader';
-import {
-  DialogFooter,
-  MiniImage,
-  StyledImage,
-} from '../population/IndividualFlexFieldPhotoModal';
+import { PhotoModal } from '../PhotoModal/PhotoModal';
 
 export interface GrievanceFlexFieldPhotoModalEditableProps {
   flexField: AllAddIndividualFieldsQuery['allAddIndividualsFieldsAttributes'][number];
@@ -32,13 +24,12 @@ export const GrievanceFlexFieldPhotoModalEditable = ({
   flexField,
 }: GrievanceFlexFieldPhotoModalEditableProps): React.ReactElement => {
   const [isEdited, setEdit] = useState(false);
-  const [turnAngle, setTurnAngle] = useState(90);
   const { id } = useParams();
   const { data } = useGrievanceTicketFlexFieldsQuery({
     variables: { id },
     fetchPolicy: 'network-only',
   });
-  const [dialogOpen, setDialogOpen] = useState(false);
+
   if (!data) {
     return null;
   }
@@ -60,39 +51,11 @@ export const GrievanceFlexFieldPhotoModalEditable = ({
           <FormikFileField field={field} form={form} />
         </Box>
       ) : (
-        <>
-          <Box display='flex' alignItems='center'>
-            <MiniImage
-              alt='photo'
-              src={picUrl}
-              onClick={() => setDialogOpen(true)}
-            />
-            <IconButton onClick={() => setEdit(true)}>
-              <CloseIcon />
-            </IconButton>
-          </Box>
-          <Dialog
-            open={dialogOpen}
-            onClose={() => setDialogOpen(false)}
-            aria-labelledby='form-dialog-title'
-          >
-            <PhotoModalHeader
-              turnAngle={turnAngle}
-              setTurnAngle={setTurnAngle}
-            />
-
-            <DialogContent>
-              <Box p={3}>
-                <StyledImage id='modalImg' alt='photo' src={picUrl} />
-              </Box>
-            </DialogContent>
-            <DialogFooter>
-              <DialogActions>
-                <Button onClick={() => setDialogOpen(false)}>CANCEL</Button>
-              </DialogActions>
-            </DialogFooter>
-          </Dialog>
-        </>
+        <PhotoModal
+          src={picUrl}
+          variant='pictureClose'
+          closeHandler={() => setEdit(true)}
+        />
       )}
     </Box>
   );
