+++ frontend/src/components/Grievances/GrievanceFlexFieldPhotoModalNewHousehold.tsx
@@ -1,17 +1,10 @@
-import { Box, Button, DialogContent } from '@material-ui/core';
-import React, { useState } from 'react';
-import { Dialog } from '../../containers/dialogs/Dialog';
-import { DialogActions } from '../../containers/dialogs/DialogActions';
+import { Box } from '@material-ui/core';
+import React from 'react';
 import {
   AllEditHouseholdFieldsQuery,
   useHouseholdFlexFieldsQuery,
 } from '../../__generated__/graphql';
-import { PhotoModalHeader } from '../PhotoModal/PhotoModalHeader';
-import {
-  DialogFooter,
-  MiniImage,
-  StyledImage,
-} from '../population/IndividualFlexFieldPhotoModal';
+import { PhotoModal } from '../PhotoModal/PhotoModal';
 
 export interface GrievanceFlexFieldPhotoModalNewHouseholdProps {
   flexField: AllEditHouseholdFieldsQuery['allEditHouseholdFieldsAttributes'][number];
@@ -22,12 +15,10 @@ export const GrievanceFlexFieldPhotoModalNewHousehold = ({
   flexField,
   householdId,
 }: GrievanceFlexFieldPhotoModalNewHouseholdProps): React.ReactElement => {
-  const [turnAngle, setTurnAngle] = useState(90);
   const { data } = useHouseholdFlexFieldsQuery({
     variables: { id: householdId },
     fetchPolicy: 'network-only',
   });
-  const [dialogOpen, setDialogOpen] = useState(false);
   if (!data) {
     return null;
   }
@@ -39,33 +30,7 @@ export const GrievanceFlexFieldPhotoModalNewHousehold = ({
   return (
     <Box style={{ height: '100%' }} display='flex' alignItems='center'>
       {picUrl ? (
-        <>
-          <MiniImage
-            alt='photo'
-            src={picUrl}
-            onClick={() => setDialogOpen(true)}
-          />
-          <Dialog
-            open={dialogOpen}
-            onClose={() => setDialogOpen(false)}
-            aria-labelledby='form-dialog-title'
-          >
-            <PhotoModalHeader
-              turnAngle={turnAngle}
-              setTurnAngle={setTurnAngle}
-            />
-            <DialogContent>
-              <Box p={3}>
-                <StyledImage id='modalImg' alt='photo' src={picUrl} />
-              </Box>
-            </DialogContent>
-            <DialogFooter>
-              <DialogActions>
-                <Button onClick={() => setDialogOpen(false)}>CANCEL</Button>
-              </DialogActions>
-            </DialogFooter>
-          </Dialog>
-        </>
+        <PhotoModal src={picUrl} />
       ) : (
         <Box style={{ height: '100%' }} display='flex' alignItems='center'>
           -
