+++ frontend/src/components/population/DocumentPopulationPhotoModal.tsx
@@ -1,18 +1,9 @@
-import { Box, Button, DialogContent } from '@material-ui/core';
-import React, { useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Dialog } from '../../containers/dialogs/Dialog';
-import { DialogActions } from '../../containers/dialogs/DialogActions';
+import React from 'react';
 import {
   IndividualNode,
-  useIndividualPhotosLazyQuery,
+  useIndividualPhotosQuery,
 } from '../../__generated__/graphql';
-import { PhotoModalHeader } from '../PhotoModal/PhotoModalHeader';
-import {
-  DialogFooter,
-  StyledImage,
-  StyledLink,
-} from './IndividualFlexFieldPhotoModal';
+import { PhotoModal } from '../PhotoModal/PhotoModal';
 
 interface DocumentPopulationPhotoModalProps {
   individual: IndividualNode;
@@ -25,10 +16,7 @@ export const DocumentPopulationPhotoModal = ({
   documentNumber,
   documentId,
 }: DocumentPopulationPhotoModalProps): React.ReactElement => {
-  const [turnAngle, setTurnAngle] = useState(90);
-  const { t } = useTranslation();
-  const [dialogOpen, setDialogOpen] = useState(false);
-  const [getPhotos, { data }] = useIndividualPhotosLazyQuery({
+  const { data } = useIndividualPhotosQuery({
     variables: { id: individual?.id },
     fetchPolicy: 'network-only',
   });
@@ -37,40 +25,11 @@ export const DocumentPopulationPhotoModal = ({
   );
 
   return (
-    <>
-      <StyledLink
-        onClick={() => {
-          setDialogOpen(true);
-          getPhotos();
-        }}
-      >
-        {documentNumber}
-      </StyledLink>
-      <Dialog
-        open={dialogOpen}
-        onClose={() => setDialogOpen(false)}
-        aria-labelledby='form-dialog-title'
-      >
-        <PhotoModalHeader
-          turnAngle={turnAngle}
-          setTurnAngle={setTurnAngle}
-          title="Document's Photo"
-        />
-        <DialogContent>
-          <Box p={3}>
-            <StyledImage
-              id='modalImg'
-              alt='document'
-              src={documentWithPhoto?.node?.photo}
-            />
-          </Box>
-        </DialogContent>
-        <DialogFooter>
-          <DialogActions>
-            <Button onClick={() => setDialogOpen(false)}>{t('CANCEL')}</Button>
-          </DialogActions>
-        </DialogFooter>
-      </Dialog>
-    </>
+    <PhotoModal
+      src={documentWithPhoto?.node?.photo}
+      linkText={documentNumber}
+      variant='link'
+      title="Document's Photo"
+    />
   );
 };
