+++ frontend/src/components/population/HouseholdFlexFieldPhotoModal.tsx
@@ -1,24 +1,15 @@
-import { Box, Button, DialogContent } from '@material-ui/core';
-import React, { useState } from 'react';
+import { Box } from '@material-ui/core';
+import React from 'react';
 import { useParams } from 'react-router-dom';
-import { Dialog } from '../../containers/dialogs/Dialog';
-import { DialogActions } from '../../containers/dialogs/DialogActions';
 import { useHouseholdFlexFieldsQuery } from '../../__generated__/graphql';
-import { PhotoModalHeader } from '../PhotoModal/PhotoModalHeader';
-import {
-  DialogFooter,
-  MiniImage,
-  StyledImage,
-} from './IndividualFlexFieldPhotoModal';
+import { PhotoModal } from '../PhotoModal/PhotoModal';
 
 export const HouseholdFlexFieldPhotoModal = ({ field }): React.ReactElement => {
   const { id } = useParams();
-  const [turnAngle, setTurnAngle] = useState(90);
   const { data } = useHouseholdFlexFieldsQuery({
     variables: { id },
     fetchPolicy: 'network-only',
   });
-  const [dialogOpen, setDialogOpen] = useState(false);
 
   if (!data) {
     return null;
@@ -28,26 +19,7 @@ export const HouseholdFlexFieldPhotoModal = ({ field }): React.ReactElement => {
   const picUrl = flexFields[field.name];
 
   return picUrl ? (
-    <>
-      <MiniImage alt='photo' src={picUrl} onClick={() => setDialogOpen(true)} />
-      <Dialog
-        open={dialogOpen}
-        onClose={() => setDialogOpen(false)}
-        aria-labelledby='form-dialog-title'
-      >
-        <PhotoModalHeader turnAngle={turnAngle} setTurnAngle={setTurnAngle} />
-        <DialogContent>
-          <Box p={3}>
-            <StyledImage id='modalImg' alt='photo' src={picUrl} />
-          </Box>
-        </DialogContent>
-        <DialogFooter>
-          <DialogActions>
-            <Button onClick={() => setDialogOpen(false)}>CANCEL</Button>
-          </DialogActions>
-        </DialogFooter>
-      </Dialog>
-    </>
+    <PhotoModal src={picUrl} />
   ) : (
     <Box display='flex' alignItems='center'>
       -
