+++ frontend/src/components/population/IndividualFlexFieldPhotoModal.tsx
@@ -1,51 +1,17 @@
-import { Box, Button, DialogContent } from '@material-ui/core';
-import React, { useState } from 'react';
-import { Link, useParams } from 'react-router-dom';
-import styled from 'styled-components';
-import { Dialog } from '../../containers/dialogs/Dialog';
-import { DialogActions } from '../../containers/dialogs/DialogActions';
+import { Box } from '@material-ui/core';
+import React from 'react';
+import { useParams } from 'react-router-dom';
 import { useIndividualFlexFieldsQuery } from '../../__generated__/graphql';
-import { PhotoModalHeader } from '../PhotoModal/PhotoModalHeader';
-
-export const DialogFooter = styled.div`
-  padding: 12px 16px;
-  margin: 0;
-  border-top: 1px solid ${({ theme }) => theme.hctPalette.lighterGray};
-  text-align: right;
-`;
-
-export const StyledImage = styled.img`
-  width: 100%;
-  height: 100%;
-  max-width: 700px;
-  max-height: 700px;
-  pointer-events: none;
-  transition: 0.4s ease-in-out;
-`;
-export const MiniImage = styled.div`
-  height: 45px;
-  width: 45px;
-  cursor: pointer;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-image: url(${({ src }) => src});
-  background-size: cover;
-`;
-
-export const StyledLink = styled(Link)`
-  color: #000;
-`;
+import { PhotoModal } from '../PhotoModal/PhotoModal';
 
 export const IndividualFlexFieldPhotoModal = ({
   field,
 }): React.ReactElement => {
   const { id } = useParams();
-  const [turnAngle, setTurnAngle] = useState(90);
   const { data } = useIndividualFlexFieldsQuery({
     variables: { id },
     fetchPolicy: 'network-only',
   });
-  const [dialogOpen, setDialogOpen] = useState(false);
 
   if (!data) {
     return null;
@@ -55,26 +21,7 @@ export const IndividualFlexFieldPhotoModal = ({
   const picUrl = flexFields[field.name];
 
   return picUrl ? (
-    <>
-      <MiniImage alt='photo' src={picUrl} onClick={() => setDialogOpen(true)} />
-      <Dialog
-        open={dialogOpen}
-        onClose={() => setDialogOpen(false)}
-        aria-labelledby='form-dialog-title'
-      >
-        <PhotoModalHeader turnAngle={turnAngle} setTurnAngle={setTurnAngle} />
-        <DialogContent>
-          <Box p={3}>
-            <StyledImage id='modalImg' alt='photo' src={picUrl} />
-          </Box>
-        </DialogContent>
-        <DialogFooter>
-          <DialogActions>
-            <Button onClick={() => setDialogOpen(false)}>CANCEL</Button>
-          </DialogActions>
-        </DialogFooter>
-      </Dialog>
-    </>
+    <PhotoModal src={picUrl} />
   ) : (
     <Box style={{ height: '100%' }} display='flex' alignItems='center'>
       -
